Bare metal build example (#1)
* Updated cmake

* First iteration of unit test on baremetal (with garbage)

* Added FetchContent of cmsis

* Refactor

* Updated description

* Updated description

* Ci test (#2)

* ci test

* Start selftests

* Unsigned int instead size_t

* Run-on changet to self-hosted

* Revert "Run-on changet to self-hosted"

This reverts commit 7e6fa9081feebacbb9c4d7531e97280b2223c12c.

* Reapply "Run-on changet to self-hosted"

This reverts commit 2818fe013de99b698b2bc8f063079cb4d56c43e3.

* Delete installs

* Added self-test on stm32f103re

* Fixe name of bin file

* Fix CI startegy

* Tricky self-test for bare metal

* CHeck lines in rtt

* Ooops...

* Optimize

+++ src/runit.c
@@ -8,5 +8,5 @@
 #include "stddef.h"
 
 char runit_at_least_one_fail = 0;
-size_t runit_counter_assert_failures = 0;
-size_t runit_counter_assert_passes = 0;
+unsigned int runit_counter_assert_failures = 0;
+unsigned int runit_counter_assert_passes = 0;
